Implement FilterEffect response setters
Add setters for both frequency response and impulse response.

+++ include/sampledata.h
@@ -5,7 +5,7 @@
 #ifndef SAMPLEDATA_H_FIPJCQHI
 #define SAMPLEDATA_H_FIPJCQHI
 
-typedef double Sample;
+typedef float Sample;
 
 class SampleData
 {
